# internal/api/v1/types.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.11.0 DO NOT EDIT.
package v1

import (
	"time"
)

const (
	JWTScopes = "JWT.Scopes"
)

// Defines values for SongService.
const (
	Unknown SongService = "unknown"
	Youtube SongService = "youtube"
)

// The object that describes a song
type Song struct {
	ArtistName   string      `json:"artist_name"`
	ArtistUrl    string      `json:"artist_url"`
	ArtworkUrl   string      `json:"artwork_url"`
	LastPlay     time.Time   `json:"last_play"`
	Playbacks    int         `json:"playbacks"`
	Service      SongService `json:"service"`
	ThumbnailUrl string      `json:"thumbnail_url"`
	Title        string      `json:"title"`
	Url          string      `json:"url"`
}

// SongService defines model for Song.Service.
type SongService string

// Error defines model for Error.
type Error struct {
	Msg string `json:"msg"`
}

// Token defines model for Token.
type Token struct {
	Token string `json:"token"`
}

// EnableMode defines model for Enable-mode.
type EnableMode struct {
	Enable bool `binding:"exists" json:"enable"`
}

// PostAuthTokenJSONBody defines parameters for PostAuthToken.
type PostAuthTokenJSONBody struct {
	// Case-insensitive
	Login string `binding:"required" json:"login"`

	// Case-sensitive
	Password string `binding:"required" json:"password"`
}

// PostMusicEnqueueServiceIdentifierJSONBody defines parameters for PostMusicEnqueueServiceIdentifier.
type PostMusicEnqueueServiceIdentifierJSONBody struct {
	Input string `binding:"required" json:"input"`
}

// PostAuthTokenJSONRequestBody defines body for PostAuthToken for application/json ContentType.
type PostAuthTokenJSONRequestBody PostAuthTokenJSONBody

// PostMusicEnqueueServiceIdentifierJSONRequestBody defines body for PostMusicEnqueueServiceIdentifier for application/json ContentType.
type PostMusicEnqueueServiceIdentifierJSONRequestBody PostMusicEnqueueServiceIdentifierJSONBody

// PostMusicLoopJSONRequestBody defines body for PostMusicLoop for application/json ContentType.
type PostMusicLoopJSONRequestBody EnableMode

// PostMusicRadioJSONRequestBody defines body for PostMusicRadio for application/json ContentType.
type PostMusicRadioJSONRequestBody EnableMode